% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{ColumnScaleSparseMatrix}
\alias{ColumnScaleSparseMatrix}
\title{Fast implementation to column scale a dgCMatrix to sum to 1}
\usage{
ColumnScaleSparseMatrix(M)
}
\arguments{
\item{M}{a dgCMatrix to be column scaled.}
}
\value{
column-scaled matrix M.
}
\description{
Fast implementation to column scale a dgCMatrix to sum to 1
}
