\name{crules-package}
\alias{crules-package}
\alias{crules}
\docType{package}
\title{
Sequential covering decision rules induction
}
\description{
Crules package provides functions for sequential covering decision rules induction, classification based on decision rules and performing cross-validation.
}
\details{
\tabular{ll}{
Package: \tab crules\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
crules package - Wojciech Malara, algorithm - Marek Sikora & Łukasz Wróbel

Maintainer: Wojciech Malara <wojcmal238@student.polsl.pl>
}
\references{
Sikora M., Wróbel Ł. (2011), Data-driven Adaptive Selection of Rules Quality Measures for Improving Rules Induction and Filtration Algorithms, Gliwice.
}
\keyword{ package }
\keyword{ machine learning }
