% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{DetectOverlappingCommunitiesSLPAw}
\alias{DetectOverlappingCommunitiesSLPAw}
\title{Pure R implementation of SLPAw (Xie et. al 2011)}
\usage{
DetectOverlappingCommunitiesSLPAw(
  W,
  num.iters = 20,
  rand.seed = 42,
  community.threshold = 0.25
)
}
\arguments{
\item{W}{a weighted walk matrix}

\item{num.iters}{a predefined number of iterations after which to terminate
the SLPA algorithm. Default 20, reported as empirically observed 
general stability threshold in the original paper.}

\item{rand.seed}{an integer seed for the RNG, default=42}

\item{community.threshold}{a number in [0,1] to determine the thresholding
value for communities to report. The function will report all
communities with a total 
number of labels >= floor(community.threshold * num.iters)
in memory for each node.}
}
\value{
a list of lists of node IDs containing communities identified
}
\description{
Pure R implementation of SLPAw (Xie et. al 2011)
}
