% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{ForecastStates}
\alias{ForecastStates}
\title{Predict cell state transition graph using data simulated using
association rules, weighted by rule confidence.}
\usage{
ForecastStates(exp.mat, rules, tau = 0.2, n.sim.steps = 5)
}
\arguments{
\item{exp.mat}{a cell (row) by gene (column) expression matrix
of class 'ngCMatrix'}

\item{rules}{a rules object}

\item{tau}{time constant for simulation}

\item{n.sim.steps}{total number of simulation steps to run.}
}
\value{
a new expression matrix E^t after `n.sim.steps` simulation
         steps with time constant tau
}
\description{
Predict cell state transition graph using data simulated using
association rules, weighted by rule confidence.
}
