% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateCellularRules}
\alias{GenerateCellularRules}
\title{Generate association rules from original data E^0}
\usage{
GenerateCellularRules(
  exp.mat,
  min.support = 0,
  min.conf = 0.25,
  max.rule.len = 4
)
}
\arguments{
\item{exp.mat}{a cell (row) by gene (column) expression matrix
of class 'ngCMatrix'}

\item{min.support}{default 0, passed to apriori algorithm}

\item{min.conf}{default 0.25, passed to apriori algorithm}

\item{max.rule.len}{default 4, passed to apriori algorithm}
}
\value{
a rules object generated from the single cell data
}
\description{
Generate association rules from original data E^0
}
