% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateIncidenceMatrix}
\alias{GenerateIncidenceMatrix}
\title{Create incidence matrix from dataset, wrapper function to allow for simpler
user interaction to various means of generating incidence matrix.}
\usage{
GenerateIncidenceMatrix(x, method = "percentile", ...)
}
\arguments{
\item{x}{a dgCMatrix of numbers (e.g. RNA counts)}

\item{method}{a key to indicate which generation method is to be used.}

\item{...}{all additional parameters will be passed to downstream functions}
}
\value{
a dgCMatrix incidence matrix
}
\description{
Create incidence matrix from dataset, wrapper function to allow for simpler
user interaction to various means of generating incidence matrix.
}
