% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateIncidenceMatrixNonzero}
\alias{GenerateIncidenceMatrixNonzero}
\title{Sets all nonzero entries to 1, creating incidence matrix.}
\usage{
GenerateIncidenceMatrixNonzero(x)
}
\arguments{
\item{x}{a dgCMatrix of numbers (e.g. RNA counts)}
}
\value{
a dgCMatrix incidence matrix
}
\description{
Sets all nonzero entries to 1, creating incidence matrix.
}
