% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateIncidenceMatrixPercentile}
\alias{GenerateIncidenceMatrixPercentile}
\title{Create incidence matrix from dataset, using column-wise percentile values to 
generate inclusion cutoffs.}
\usage{
GenerateIncidenceMatrixPercentile(x, percentile.cutoff = 0.8)
}
\arguments{
\item{x}{a dgCMatrix of numbers (e.g. RNA counts)}

\item{percentile.cutoff}{the percentile at which 
to consider "prescence" in the incidence matrix.
Default = 0.8.}
}
\value{
a dgCMatrix incidence matrix
}
\description{
Create incidence matrix from dataset, using column-wise percentile values to 
generate inclusion cutoffs.
}
