% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateMarkovChain}
\alias{GenerateMarkovChain}
\title{Calculate new id vector (sigma)}
\usage{
GenerateMarkovChain(ids, exp.mat, x.prime)
}
\arguments{
\item{ids}{a vector of cell cluster ids to generate 
transition probabilities}

\item{exp.mat}{original expression data (E^0)}

\item{x.prime}{predicted expression data (E^t)}
}
\value{
vector of new ids
}
\description{
Calculate new id vector (sigma)
}
