% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateMarkovTrace}
\alias{GenerateMarkovTrace}
\title{Generate Markov trace for simulation of `exp.mat` on the provided set of `rules`
with time constant tau and for n.sim.steps steps.}
\usage{
GenerateMarkovTrace(ids, exp.mat, rules, tau = 0.2, n.sim.steps = 5)
}
\arguments{
\item{ids}{a vector of cell identities (e.g. cluster ids, cell types)}

\item{exp.mat}{a cell (row) by gene (column) expression matrix
of class 'ngCMatrix'}

\item{rules}{a rules object}

\item{tau}{time constant for simulation}

\item{n.sim.steps}{total number of simulation steps to run.}
}
\value{
a (cell x sim step) matrix capturing cell states at each step of the
         simulation for fitting to a markov model.
}
\description{
Generate Markov trace for simulation of `exp.mat` on the provided set of `rules`
with time constant tau and for n.sim.steps steps.
}
