% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateWalkMatrix}
\alias{GenerateWalkMatrix}
\title{Creates a walk matrix from a set of rules.}
\usage{
GenerateWalkMatrix(rules, tau = 0.2)
}
\arguments{
\item{rules}{a rules object}

\item{tau}{defualt 0.2. Time constant for the walk matrix, in [0,1].}
}
\value{
a gene x gene walk matrix W.
}
\description{
Creates a walk matrix from a set of rules.
}
