% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{Identity}
\alias{Identity}
\title{Identity function to test package configuration}
\usage{
Identity(dummy)
}
\arguments{
\item{dummy}{a value to return}
}
\value{
the dummy value
}
\description{
Identity function to test package configuration
}
