% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{ListenerModeOrRandom}
\alias{ListenerModeOrRandom}
\title{Internal LISTENER function to be used in SLPAw algorithm.}
\usage{
ListenerModeOrRandom(x, ...)
}
\arguments{
\item{x}{a vector of values}

\item{...}{dummy additional arguments that may be passed to other listeners}
}
\value{
the mode or weightd mode of provided values, or a random number 
        amongst tied modes.
}
\description{
Internal LISTENER function to be used in SLPAw algorithm.
}
