% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{ListenerProbWeighted}
\alias{ListenerProbWeighted}
\title{Internal LISTENER function to be used in SLPAw algorithm.}
\usage{
ListenerProbWeighted(x, weights, ...)
}
\arguments{
\item{x}{a vector of values}

\item{weights}{an optional set of weights to accompany the values}

\item{...}{dummy additional arguments that may be passed to other listeners}
}
\value{
the mode or weightd mode of provided values, or a random number 
        amongst tied modes.
}
\description{
Internal LISTENER function to be used in SLPAw algorithm.
}
