% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{NewValsFromPVector}
\alias{NewValsFromPVector}
\title{ngCmatric-specific manipulation, internal helper function
to weight values of matrix by a "confidence" level obtained
from the association rule learning stage. Confidence may be
replaced with any other metric on the rule space}
\usage{
NewValsFromPVector(p, vals)
}
\arguments{
\item{p}{the p vector from an ngCMatrix}

\item{vals}{the vector of values to replace per column:
e.g. confidence}
}
\value{
a vector which may be assigned to the x slot of 
         the ngCMatrix used for parameter "p".
}
\description{
ngCmatric-specific manipulation, internal helper function
to weight values of matrix by a "confidence" level obtained
from the association rule learning stage. Confidence may be
replaced with any other metric on the rule space
}
