% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{RankRulesHypergraph}
\alias{RankRulesHypergraph}
\title{Generate representation of rule hypergraph from which to derive
interestingness measure for rules in the style of CONFIGV
(Santolucito et. al, OOPSLA 2017)}
\usage{
RankRulesHypergraph(rules)
}
\arguments{
\item{rules}{a rules output object from arules apriori or eclat}
}
\value{
a vector of length equal to the number of rules specifying 
        a rank score by hypergraph degree as defined in
        (Santolucito et. al, OOPSLA 2017)
}
\description{
Generate representation of rule hypergraph from which to derive
interestingness measure for rules in the style of CONFIGV
(Santolucito et. al, OOPSLA 2017)
}
