% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{RowScaleSparseMatrix}
\alias{RowScaleSparseMatrix}
\title{Fast implementation to row scale a dgCMatrix to sum to 1}
\usage{
RowScaleSparseMatrix(M)
}
\arguments{
\item{M}{a dgCMatrix to be row scaled.}
}
\value{
column-scaled matrix M.
}
\description{
Fast implementation to row scale a dgCMatrix to sum to 1
}
