% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{TrimSmallComms}
\alias{TrimSmallComms}
\title{Small helper function to trim small communities from list of lists}
\usage{
TrimSmallComms(communities, min.community.size = 2)
}
\arguments{
\item{communities}{a vector of values}

\item{min.community.size}{an optional set of weights to accompany the values}
}
\value{
a pruned list of lists with only those of size greater than
        `min.community.size`.
}
\description{
Small helper function to trim small communities from list of lists
}
