% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{WalkCellState}
\alias{WalkCellState}
\title{Simulate the cell states using a walk matrix generated
from the association rules, re-normalizing the matrix at
each step.}
\usage{
WalkCellState(x, W, nsteps = 5)
}
\arguments{
\item{x}{a cell (row) by gene (column) ngCMatrix of expression
data to perform a walk along}

\item{W}{the row-normalized walk matrix}

\item{nsteps}{the total number of time steps to run the simulation}
}
\value{
a new x' of the original matrix after nsteps iterations.
}
\description{
Simulate the cell states using a walk matrix generated
from the association rules, re-normalizing the matrix at
each step.
}
