% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{CalcCommPenaltyFromGMT}
\alias{CalcCommPenaltyFromGMT}
\title{Compare communities list of lists to *.gmt symbol table from MSigDB and
generate a concordance score based on the Jaccard dissimilarity index.}
\usage{
CalcCommPenaltyFromGMT(communities, gmt.filename)
}
\arguments{
\item{communities}{a list of lists containing symbols for gene communities}

\item{gmt.filename}{a filename for gmt against which to test communities}
}
\value{
a dataframe containing the penalty weights for each community,
        as well as the name of the closest gene set in the provided GMT.
}
\description{
Compare communities list of lists to *.gmt symbol table from MSigDB and
generate a concordance score based on the Jaccard dissimilarity index.
}
