% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crules.R
\name{GenerateCellularRules}
\alias{GenerateCellularRules}
\title{Generate association rules from original data E^0}
\usage{
GenerateCellularRules(
  exp.mat,
  arm.algorithm = "apriori",
  supp = 0,
  conf = 40,
  ...
)
}
\arguments{
\item{exp.mat}{a cell (row) by gene (column) expression matrix
of class 'ngCMatrix'}

\item{arm.algorithm}{algorithm used for frequent itemset mining,
supports 'apriori' (default) and 'eclat'.}

\item{supp}{Default 0, minimum support to pass to arm.algorithm in [0,100]}

\item{conf}{Default 40, minimum confidence to pass to arm.algorithm in [0,100]}

\item{...}{additional parameters to pass for rule mining}
}
\value{
a rules object generated from the single cell data
}
\description{
Generate association rules from original data E^0
}
